操作系统赛/requirements.txt
PyYAML>=5.4
argparse
